// src/common/schemas.types.ts
import { z } from "astro:schema";

export const coordinateSchema = z.object({
    x: z.number(),
    y: z.number(),
});

export const customLocationSchema = z.object({
    address: z.string(),
    delivery: z.string(),
    description: z.string(),
    category: z.string(),
    city: z.string(),
    google_id: z.string(),
    latitude: z.number(),
    longitude: z.number(),
    main_image: z.string(),
    open_time_friday: z.string().optional(),
    open_time_monday: z.string().optional(),
    open_time_saturday: z.string().optional(),
    open_time_sunday: z.string().optional(),
    open_time_thursday: z.string().optional(),
    open_time_tuesday: z.string().optional(),
    open_time_wednesday: z.string().optional(),
    page_url: z.string(),
    phone: z.string(),
    place_id: z.string(),
    plus_code: z.string(),
    price_range: z.string(),
    restaurant_name: z.string(),
    review_points: z.number(),
    reviews: z.number(),
    state: z.string(),
    website: z.string(),
    zipcode: z.number(),
});

export const customLocation = {
    address: "" as string,
    delivery: "" as string,
    description: "" as string,
    category: "" as string,
    city: "" as string,
    google_id: "" as string,
    latitude: 0 as number,
    longitude: 0 as number,
    main_image: "" as string,
    open_time_friday: "" as string | undefined,
    open_time_monday: "" as string | undefined,
    open_time_saturday: "" as string | undefined,
    open_time_sunday: "" as string | undefined,
    open_time_thursday: "" as string | undefined,
    open_time_tuesday: "" as string | undefined,
    open_time_wednesday: "" as string | undefined,
    page_url: "" as string,
    phone: "" as string,
    place_id: "" as string,
    plus_code: "" as string,
    price_range: "" as string,
    restaurant_name: "" as string,
    review_points: 0 as number,
    reviews: 0 as number,
    state: "" as string,
    website: "" as string,
    zipcode: 0 as number,
};

export const googleLocationSchema = z.object({
    address: z.string(),
    coord: coordinateSchema,
    id: z.string(),
    hours_1_sunday_open: z.boolean(),
    hours_2_monday_open: z.boolean(),
    hours_3_tuesday_open: z.boolean(),
    hours_4_wednesday_open: z.boolean(),
    hours_5_thursday_open: z.boolean(),
    hours_6_friday_open: z.boolean(),
    hours_7_saturday_open: z.boolean(),
    name: z.string(),
    phone: z.string(),
    website: z.string(),
});